update dbc structures

Conflicts:
	src/game/Creature.cpp
	src/game/CreatureAI.cpp
	src/game/DBCEnums.h
	src/game/DBCStores.cpp
	src/game/DBCStructure.cpp
	src/game/DBCStructure.h
	src/game/DBCfmt.h
	src/game/ObjectMgr.cpp
	src/game/Spell.cpp
	src/game/SpellAuras.cpp
	src/game/SpellEffects.cpp
	src/game/SpellMgr.cpp
	src/game/SpellMgr.h
	src/game/Totem.cpp
	src/game/Unit.cpp
	src/game/UnitAuraProcHandler.cpp

+++ src/game/BattleGround/BattleGroundMgr.cpp
@@ -1052,7 +1052,7 @@ void BattleGroundQueue::Update(BattleGroundTypeId bgTypeId, BattleGroundBracketI
 
 BattleGroundMgr::BattleGroundMgr() : m_ArenaTesting(false)
 {
-    for (uint8 i = BATTLEGROUND_TYPE_NONE; i < MAX_BATTLEGROUND_TYPE_ID; ++i)
+    for (uint32 i = BATTLEGROUND_TYPE_NONE; i < MAX_BATTLEGROUND_TYPE_ID; ++i)
         m_BattleGrounds[i].clear();
     m_NextRatingDiscardUpdate = GetRatingDiscardTimer();
     m_Testing = false;
